Move get_swaps() into rfc003 module
The get swaps endpoint is tied to rfc003 by the fact that it uses the
function `build_rfc003_siren_entity()`.  This function uses the
`with_swap_types` macro, this is never going to work with the new split
protocols. `get_swaps()` and its handler should be over in the rfc003
module along with the other rfc003 endpoint handlers to make this
explicit.
