--- cnd/src/http_api/routes/index/handlers.rs
@@ -1,3 +1,0 @@
-mod get_swaps;
-
-pub use self::get_swaps::handle_get_swaps;
