+++ cnd/src/http_api/routes/rfc003/handlers.rs
@@ -1,9 +1,11 @@
 mod action;
 mod get_swap;
+mod get_swaps;
 pub mod post_swap;
 
 pub use self::{
     action::{handle_action, InvalidAction, InvalidActionInvocation},
     get_swap::handle_get_swap,
+    get_swaps::handle_get_swaps,
     post_swap::handle_post_swap,
 };
